requirements.txt
dbt-core==1.3.0
dbt-bigquery==1.3.0
pyspark==3.3.1
google-cloud-storage
google-cloud-bigquery
apache-airflow-providers-apache-spark
requests

